Some more refactor and many bugfixes.
